+++ src/config/database.js
@@ -10,6 +10,15 @@ const sequelize = new Sequelize(env.db.name, env.db.user, env.db.password, {
     underscored: true,
     timestamps: true,
   },
+  pool: {
+    max: 10,
+    min: 0,
+    acquire: 60000,
+    idle: 10000,
+  },
+  dialectOptions: {
+    connectTimeout: 60000,
+  },
 });
 
 module.exports = sequelize;